--- nest_lesson_3/src/modules/usersModule/users/models/users.input.model.ts
@@ -1,8 +1,0 @@
-import { EmailConfirmation } from '../domain/user.entity';
-
-export class CreateUserInput {
-  login: string;
-  email: string;
-  password: string;
-  emailConfirmation?: EmailConfirmation;
-}
