--- nest_lesson_3/src/modules/usersModule/users/models/users.output.model.ts
@@ -1,14 +1,0 @@
-import { UserDocument } from '../domain/user.entity';
-
-export const userOutputModelMapper = (user: UserDocument) => {
-  if (!user) {
-    return null;
-  }
-
-  return {
-    id: user.id,
-    login: user.login,
-    email: user.email,
-    createdAt: user.createdAt,
-  };
-};
